Added the search and caching for products.

+++ database/migrations/2018_01_18_214131_create_products_table.php
@@ -15,7 +15,7 @@ class CreateProductsTable extends Migration
     {
         Schema::create('products', function (Blueprint $table) {
             $table->increments('id');
-            $table->integer('productId');
+            $table->bigInteger('productId');
             $table->string('title');
             $table->text('description');
             $table->string('vendor');
